forgot to add this bookHW4 and note that file2.dat will have a zero
because file1.dat does not exist and the program does not check that
opening this file failed like infile.fail() and an if statement. Thus,
either the int someInteger already has a zero in it because I gave it
nothing at declaration or the >> extraction operator or << is adding a zero if
variable is null. I do not know at this time.